Update patch set 11
Patch Set 11:

This did not resolve the introspection timeouts in the HP case that I tested. I don't have a major issue with accepting this if it works in some case.
We need to be clear on when this is to be used though

Patch-set: 11
Reviewer: Ronelle Landy <1000779@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0