Add disk-size-hint to prep-images repo
Previously this code lived in prep-baremetal repo.
Since this repo is called often before introspection
is run, the disk0-size-hints are called  before
they can impact the correct values.

disk-size-hints is historically used only on some
baremetal environments but can be included in
validation tests on virt environments as well.

Change-Id: I3c0fe67d2b4adcafa65057a8536e2629c62491e7