Update patch set 1
Patch Set 1: Code-Review+2

This seems reasonable for the short term... once we get all the roles in one place we need to rethink our various control mechanisms.

I also think it might be better to just have different templates for multinode vs not in this role as it seems like we have pretty mutually exclusive steps.

In any case lets not block on that now.

Patch-set: 1
Label: Code-Review=+2